temp_requirements.txt
edge_tts
fastapi
faster_whisper
pydantic
pymongo
python-dotenv
langchain
langchain-openai
pypdf
tiktoken
langchain-huggingface
langchain-chroma
python-multipart
chromadb
fitz
numpy
pdfplumber
sentence_transformers